pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "magpack"
version = "0.2.1"
license = "BSD-3-Clause"
authors = [
  { name="Andreas Apseros"},
]
description = "Tools for analysing 3D magnetic simulations and reconstructions"
readme = "README.rst"
requires-python = ">=3.9"
dependencies = [
    "numpy","pyevtk", "matplotlib", "pillow", "scipy"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/apseros/magpack"
Issues = "https://github.com/apseros/magpack/issues"